Conflicts:
	conf/schema/natural.json
	conf/schema/public_transport.json

+++ hoot-core/src/main/cpp/hoot/core/io/PbfReader.h
@@ -244,7 +244,7 @@ private:
 
   double _convertLat(long lat);
 
-  ElementId _convertToElementId(int id, int memberType);
+  ElementId _convertToElementId(long id, int memberType);
 
   long _createRelationId(long fromFile);
